Support BigDecimal input to OvertimeSpecification to be able to pass value from ReportOvertimeDto directly - this will provide a support to decimal parts of reported overtime.

+++ server/src/main/java/com/urbanpawel/overtime/overtime/endpoint/OvertimeEndpoint.java
@@ -20,7 +20,7 @@ class OvertimeEndpoint {
 
     @PostMapping
     public ResponseEntity reportOvertime(@RequestBody ReportOvertimeDto request) {
-        overtimeService.forDate(request.date()).hours(request.hours().intValue()).save();
+        overtimeService.forDate(request.date()).hours(request.hours()).save();
         return ResponseEntity.noContent().build();
     }
